feat: add exec() to run builds
